# src/frontend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Fitness Tracking API
  version: 1.0.0
  description: |
    A comprehensive fitness tracking API with voice integration, workout management, 
    exercise libraries, and LLM-powered workout generation capabilities.
  
    ## Features
    - User authentication with JWT tokens
    - Workout creation and tracking
    - Exercise library with media support
    - Voice-powered workout logging
    - LLM workout generation
    - Goal setting and progress tracking
    - Offline support with sync capabilities
    
  contact:
    name: API Support
    email: support@fitnessapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.fitnessapp.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /workouts/{id}/sets:
    post:
      tags:
        - Workouts
      summary: Add set to workout
      description: Add a new set to an existing workout
      parameters:
        - $ref: '#/components/parameters/WorkoutId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutSetCreateRequest'
      responses:
        '201':
          description: Set added successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkoutSet'
  /sets/{id}:
    patch:
      tags:
        - Workouts
      summary: Update workout set
      description: Update an existing workout set
      parameters:
        - $ref: '#/components/parameters/SetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutSetUpdateRequest'
      responses:
        '200':
          description: Set updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkoutSet'

    delete:
      tags:
        - Workouts
      summary: Delete workout set
      description: Delete a workout set
      parameters:
        - $ref: '#/components/parameters/SetId'
      responses:
        '200':
          description: Set deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    WorkoutId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Workout ID
      example: "workout_123"
...
    SetId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Set ID
      example: "set_789"
...
  schemas:
...
    WorkoutSet:
      type: object
      required:
        - id
        - exerciseId
        - setNumber
        - completed
      properties:
        id:
          type: string
          example: "set_123"
        exerciseId:
          type: string
          example: "exercise_bench_press"
        setNumber:
          type: integer
          minimum: 1
          example: 1
        reps:
          type: integer
          minimum: 0
          example: 8
        weight:
          type: number
          minimum: 0
          example: 135.5
        time:
          type: integer
          minimum: 0
          description: Duration in seconds
          example: 30
        distance:
          type: number
          minimum: 0
          description: Distance in meters
          example: 1000
        rpe:
          type: integer
          minimum: 1
          maximum: 10
          description: Rate of Perceived Exertion (1-10)
          example: 7
        notes:
          type: string
          example: "Felt good, could do more"
        completed:
          type: boolean
          example: true
        completedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
...
    WorkoutSetCreateRequest:
      type: object
      required:
        - exerciseId
        - setNumber
      properties:
        exerciseId:
          type: string
          example: "exercise_bench_press"
        setNumber:
          type: integer
          minimum: 1
          example: 1
        reps:
          type: integer
          minimum: 0
          example: 8
        weight:
          type: number
          minimum: 0
          example: 135.5
        time:
          type: integer
          minimum: 0
        distance:
          type: number
          minimum: 0
        rpe:
          type: integer
          minimum: 1
          maximum: 10
        notes:
          type: string
    WorkoutSetUpdateRequest:
      type: object
      properties:
        reps:
          type: integer
          minimum: 0
        weight:
          type: number
          minimum: 0
        time:
          type: integer
          minimum: 0
        distance:
          type: number
          minimum: 0
        rpe:
          type: integer
          minimum: 1
          maximum: 10
        notes:
          type: string
        completed:
          type: boolean
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
...